requirements.txt
alembic==1.4.2
appnope==0.1.0
argon2-cffi==20.1.0
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.3
Babel==2.8.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
bleach==3.1.5
blinker==1.4
bs4==0.0.1
cchardet==2.1.5
certifi==2020.6.20
cffi==1.13.2
chardet==3.0.4
click==7.1.2
clickhouse-driver==0.1.4
clickhouse-sqlalchemy==0.1.4
console-progressbar==1.1.2
cryptography==2.8
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dnspython==1.16.0
docutils==0.16
dominate==2.5.1
email-validator==1.1.1
eml-parser==1.11.5
entrypoints==0.3
et-xmlfile==1.0.1
facebook-sdk==3.1.0
flake8==3.8.3
Flask==1.1.2
Flask-Admin==1.5.6
Flask-BabelEx==0.9.4
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
googletrans==2.4.0
gunicorn==20.0.4
idna==2.10
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
jwt==1.0.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
libmagic==1.0
lxml==4.5.2
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.0
numpy==1.19.1
openpyxl==3.0.3
packaging==20.4
pandas==1.1.1
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.1
passlib==1.7.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.19
pydocstyle==5.0.2
pyflakes==2.2.0
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.3
pymongo==3.9.0
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-jsonrpc-server==0.3.4
python-language-server==0.34.1
pytz==2020.1
pyzmq==19.0.2
qtconsole==4.7.5
QtPy==1.9.0
redis==3.5.3
requests==2.24.0
rope==0.17.0
rq==1.5.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.0.1
speaklater==1.3
SQLAlchemy==1.3.18
sshtunnel==0.1.5
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
transliterate==1.10.2
tzlocal==2.1
ujson==1.35
urllib3==1.25.9
visitor==0.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.3.1
xlrd==1.2.0
XlsxWriter==1.2.8
yapf==0.30.0
